+++ packages/codegen/src/codegen/graphql.ts
@@ -1,13 +1,13 @@
 import { type CodegenConfig, generate } from "@graphql-codegen/cli";
 import { TypeScriptPluginConfig } from "@graphql-codegen/typescript";
-import { readdirSync } from "fs";
-
+import { plugin } from "@acaldas/graphql-codegen-typescript-validation-schema";
+import { readdirSync } from "node:fs";
 const getDirectories = (source: string) =>
   readdirSync(source, { withFileTypes: true })
     .filter((dirent) => dirent.isDirectory())
     .map((dirent) => dirent.name);
 
-const tsConfig: TypeScriptPluginConfig = {
+export const tsConfig: TypeScriptPluginConfig = {
   strictScalars: true,
   scalars: {
     Unknown: "unknown",
@@ -25,7 +25,31 @@ const tsConfig: TypeScriptPluginConfig = {
   inputMaybeValue: "T | null | undefined",
 };
 
-function schemaConfig(name: string, dir: string): CodegenConfig["generates"] {
+export type ValidationSchemaConfigType = Parameters<typeof plugin>[2];
+
+export const validationConfig: ValidationSchemaConfigType = {
+  importFrom: `./types`,
+  schema: "zod",
+  ...tsConfig,
+  scalarSchemas: {
+    Unknown: "z.unknown()",
+    DateTime: "z.string().datetime()",
+    Attachment: "z.string()",
+    Address:
+      "z.custom<`${string}:0x${string}`>((val) => /^[a-zA-Z0-9]+:0x[a-fA-F0-9]{40}$/.test(val as string))",
+  },
+  directives: {
+    equals: {
+      value: ["regex", "/^$1$/"],
+    },
+  },
+  withObjectType: true,
+};
+
+export function schemaConfig(
+  name: string,
+  dir: string,
+): CodegenConfig["generates"] {
   return {
     [`${dir}/${name}/gen/schema/types.ts`]: {
       schema: [
@@ -41,24 +65,7 @@ function schemaConfig(name: string, dir: string): CodegenConfig["generates"] {
     [`${dir}/${name}/gen/schema/zod.ts`]: {
       schema: `${dir}/${name}/schema.graphql`,
       plugins: ["@acaldas/graphql-codegen-typescript-validation-schema"],
-      config: {
-        importFrom: `./types`,
-        schema: "zod",
-        ...tsConfig,
-        scalarSchemas: {
-          Unknown: "z.unknown()",
-          DateTime: "z.string().datetime()",
-          Attachment: "z.string()",
-          Address:
-            "z.custom<`${string}:0x${string}`>((val) => /^[a-zA-Z0-9]+:0x[a-fA-F0-9]{40}$/.test(val as string))",
-        },
-        directives: {
-          equals: {
-            value: ["regex", "/^$1$/"],
-          },
-        },
-        withObjectType: true,
-      },
+      config: validationConfig,
     },
   };
 }