+++ packages/codegen/src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from "./cli";
 export * from "./mock";
+export { init as createProject } from "../create-lib/init";
+export { getPackageManager } from "../create-lib/command";